Replaced 3 tools with one "kripke" and replaced input-deck with command line parser

+++ src/Kripke/Timing.cpp
@@ -72,7 +72,7 @@ void Timing::print(void) const {
   printf("\nTimers:\n");
   printf("  %-16s  %12s  %12s\n", "Timer", "Count", "Seconds");
   for(int i = 0;i < names.size();++ i){
-    printf("  %-16s  %12d  %12.5lf\n", names[i].c_str(), ord_timers[i]->count, ord_timers[i]->total_time);
+    printf("  %-16s  %12d  %12.5lf\n", names[i].c_str(), (int)ord_timers[i]->count, ord_timers[i]->total_time);
   }
 }
